1、修改段10过渡出错问题 2、经客户上机验证，步进电机工作模式正常 3、目前客户发现存在问题，（1)RS485在接通开机信号后，通讯连接失败（2）产能数据丢失

+++ src/stm32f4_rtos/User/product_cal.h
@@ -98,7 +98,7 @@ u32 product_uncomplete_meter(PRODUCT_PARA para,PEILIAO_PARA peiliao);
 u32 count_per_kilo(u32 pluse);
 u32 product_complete_kilo(PRODUCT_PARA para,PEILIAO_PARA peiliao);
 u32 product_uncomplete_kilo(PRODUCT_PARA para,PEILIAO_PARA peiliao);
-float get_float_1bit(float data);
+u16 get_float_1bit(float data);
 u8 get_class_time(RTC_TIME *time,DEVICE_INFO *para);
 u8 is_same_data(u32 card,u32 *buf_lib,u16 buf_len);
 u8 get_card_type(u32 id);
